Clean up Getting Started tutorial; add "Deploy to Futurenet" (#516)
* feat: move Freighter and Running RPC to Reference

We want to keep the stuff in Getting Started hyper-focused on a quick
end-to-end tour of Soroban. Now that there's a public RPC endpoint, it
is much more straightforward to have people deploy contract right to it,
rather than running their own nodes. Running your own node is a more
advanced concept that dabblers need not muck with.

So this moves both the "Deploy to a Local Network" and "Deploy to
Futurenet" pages to the RPC page, in the Reference section. This
required some reorganization & rework of the concepts to make it fit the
new context. I also updated some instructions, such as teaching
people to use `soroban config network` and `soroban config identity`,
which were not included in the old "Deploy to…" pages.

This commit also moves the "Connect Freighter Wallet" to Reference,
since it's a bit in-the-weeds, which is a good feature for a Reference
document but not for an end-to-end Tutorial.

Coming Soon™: a new "Deploy Contracts" step in the Getting Started
section, which uses the public RPC Futurenet endpoint. This will be
followed by a "Create a Dapp" step, which will use the new JS Lib
generation built into the CLI, and go over relevant Freighter details.

* fix: clean up Hello World walk-through

- explicitly instruct people to call it `hello-soroban`
- move some single-use footnote-style links inline, to make them easier
  to see/understand/update later
- remove spacing around tips for easier visual parsing
- remove "Hello World Example" section, since this IS the Hello World
  example
- remove unused & unneeded footnote-style link definitions
- clean up example code so it doesn't give warnings on copy/paste

* feat(getting-started): clean up Storing Data

I'm working toward having people create a frontend Dapp with multiple
contracts in a `contracts` folder, so I added a whole part at the
beginning walking people through setting up a Cargo Workspace.

This page also had more information than necessary about State
Expiration; I moved this to the State Expiration doc, which had been
practically blank. I reorganized this information somewhat; most
notably, the "Best Practices" section had been practically empty. I
moved some miscellaneous information to it; it should maybe be renamed
"more information about state expiration" rather than "best practices",
since it DOESN'T currently contain best practices, and I am not
qualified to say what Best Practices might actually be. (Note that in the
old version, the "Best Practices" section was blank! It was a misplaced
heading, with nothing relevant below.)

* fix: add missing "be"

Co-authored-by: Willem Wyndham <wyndham@cs.unc.edu>

* feat: add 3. Deploy to Futurenet

This contains the relevant details people will probably need to know
when first getting started with Soroban.

* fix: run prettier

* feat: commit, summarize, save IDs to files

Instructing people to commit to git will help them think about their
work more clearly, and will also help them avoid mistakes and get better
help when they need it.

I also added summaries to each of these tutorial pages, to review what
was learned and set up the next lesson.

Finally, I decided to have them save contract IDs to files in their
`.soroban` folder, rather than in more-ephemeral environment variables.
This will make it easier to have them use `soroban contract bindings` in
the "Build a Dapp" step, since they'll be able to reference these files
in their package.json scripts.

* fix: run lint

---------

Co-authored-by: Willem Wyndham <wyndham@cs.unc.edu>