Update smart-contract-deployment.mdx (#645)
The it is an issue in the vault contract's withdraw function which will incorrectly update the value of Reserve.

In this line

    let balance = get_token_balance(&e);

balance is updated AFTER the transfer of the user deposit + yield has already occurred. 

Thus the balance fetched is already the current "reserve" balance of the Vault. 

So when the following line of code executes:

put_reserve(&e, balance - amount);

the Reserve is updated to the balance of the Vault contract - the amount withdrawn by the user, so it will count the withdrawn 2 times. 

Indeed when following the test the result from "get_rsrvs" in the first scenario will be :

"-1" 

Initial balance of the Vault: 200 - amount: 100 + yield: 1 (from the transfer) - amount: 100 (from the put_reserve input). 

For further improvements the get_rsrvs should only account for the amount of reward tokens that the vault is holding which can be useful to obtain the deposited amount (vault balance - reserves).

Additionally the --wasm argument in the "invoke" command is not needed and it was removed.