Create Terminal-UI-Guide.mdx
This commit adds a guide for the Comet Liquidity Pool Dapp Terminal User Interface (TUI) project. The project provides a comprehensive guide on building a functional user interface for decentralized applications using the inquirer package in Node.js. The primary focus is on adding liquidity to a liquidity pool contract, specifically the 'comet' contract, with the hope that the guide provides help for someone to setup a TUI for a different dapp.