move stellar-cli man page out of dev-tools directory
Having it nested under the "Developer Tools" sidebar category is
confusing in its current state, and the _kind_ of dev tools we
list in that section is very different from the stellar-cli man
page content. Having its own page in the sidebar feels more
intuitive and like a better way to promote its access.