update the translations fix script to include MDX comments
Crowdin does some kind of "processing" or "parsing" to MDX files,
when we use the `crowdin download` command. This replaces any
occurence of MDX comments like `{/* ... */}` with `{/_ ... _/}`.
I think this is because it assumes the single-`*` asterisks are
meant to denote underline syntax, so it replaces them with the
configured underline character, `_`.

Anyway, this changes our fixup script to account for that. It
uses find and xargs to replace any instances of those errant
non-comment strings, and replaces them with valid MDX comments,
resulting in builds that actually build.