💱 Simply Liquidity Horizon Documentation This cleans up the wording around #1061 without reviewing the pages we notify users about `consistency`.
I think a better long-term solution is a "Consistency" flag added to the affected endpoint pages. This could be a simple button akin to the "Supports Streaming" info box. However, even that seems a verbose solution which would also do well to exist as a simple button (akin to  ledgerStorage) > two sentences.