feat: more approachable & in-depth indexer intro
Clean up the Data section's sidebar and add MUCH more info & context to
the "Data -> Indexers" page.

While mostly adding info, this removes the small Goldsky
tutorial+screenshots that had been on this page. I'm not sure our
documentation is the right place to walk through a partner's solution in
this way. We already link out to Goldsky's documentation, which is quite
good, containing more & better screenshots than what we had.
Additionally, this sort of tutorial-via-screenshots content did not fit
as well within the reworked Indexers homepage, which is a bit
higher-level. I wasn't sure it was worth keeping this content anywhere,
and decided to remove it for now.

Co-authored-by: Elliot Voris <elliot@voris.me>