Horizon ain't deprecated yet
"deprecated", to my engineering-trained mind, means "there is a better
alternative available that you should now switch to."

There is ***no*** viable alternative to Horizon today. We therefore
cannot declare it to be deprecated.

When I read that it is _already_ deprecated, it makes me think I can
avoid using it in a project. If I cannot realistically avoid using it,
at least not without great cost in both engineering research time and
infrastructure bills, then the assertion that it is already deprecated
will cause me distress.

Once we have good Portfolio API-style indexing services in place, with
freemium models appropriate for the whole gamut of
hackathon-startups-to-enterprises, then we can consider Horizon to be
deprecated. Until then, there's just no other reasonable way to get a
whole host of necessary data, such as token balances and circulating
supply.

This updates the "Horizon is deprecated" warning to say that it _will_
be deprecated in the future, deep-linking to the "Portfolio APIs"
section of the [newly-revamped indexers page][1].

  [1]: https://github.com/stellar/stellar-docs/pull/2055