Chore: cleanup components imports (#2132)
* remove `Alert` imports where they're not used

* remove the `CodeExample` import everywhere

we've added it to the list of globally available components so it's
not necessary to import it into every page it's used on anymore

* add a note to the README file about the code example component

* remove `AttributeTable`, `EndpointsTable`, `ExampleResponse` imports

same reason as the `CodeExample` component, these are globally
available, and don't need to be re-imported every time.

* fix formatting for horizon's create claimable balance object

* fixup some wallet SDK imports

* remove custom component that's only used in one place

* remove the seldom-used `Alert` component

it's a wrapper for only one type of admonition, and doesn't really
seem all that useful. people tend to understand and use `:::info`
without much trouble, so the lower-overhead seems like a benefit